requirements.txt
asgiref==3.8.1
certifi==2025.6.15
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
charset-normalizer==3.4.2
colorama==0.4.6
contourpy==1.3.2
cryptography==45.0.4
cycler==0.12.1
Django==5.2.3
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
fonttools==4.58.4
idna==3.10
inflection==0.5.1
kiwisolver==1.4.8
matplotlib==3.10.3
msgpack==1.1.1
mysqlclient>=2.1.1
mysql-connector-python==8.3.0
Faker==25.0.0
numpy==2.3.1
opencv-python==4.11.0.86
packaging==25.0
pandas==2.3.0
pillow==11.2.1
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.9.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
qrcode==8.2
redis==6.2.0
reportlab==4.4.2
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
whitenoise==6.9.0
